Report latency as floats, with units of seconds
[#128257205]

+++ src/example-apps/proxy/handlers/stats_handler.go
@@ -8,11 +8,11 @@ import (
 
 type Stats struct {
 	// Locker  sync.Locker
-	Latency []int `json:"latency"`
+	Latency []float64 `json:"latency"`
 	sync.RWMutex
 }
 
-func (s *Stats) Add(latency int) {
+func (s *Stats) Add(latency float64) {
 	s.Lock()
 	defer s.Unlock()
 	s.Latency = append(s.Latency, latency)
@@ -21,10 +21,10 @@ func (s *Stats) Add(latency int) {
 func (s *Stats) Clear() {
 	s.Lock()
 	defer s.Unlock()
-	s.Latency = []int{}
+	s.Latency = []float64{}
 }
 
-func (s *Stats) GetLatency() []int {
+func (s *Stats) GetLatency() []float64 {
 	s.RLock()
 	defer s.RUnlock()
 	return s.Latency
